// src/models/tmdb/enum.ts
export enum MediaType {
  tv = "tv",
  movie = "movie",
  person = "person",
}

export enum ImageType {
  logo = "logo",
  backdrop = "backdrop",
  poster = "poster",
  profile = "profile",
  still = "still",
}

export enum TimeTrending {
  day = "day",
  week = "week",
}

export enum SortImdb {
  popularityAsc = "popularity.asc",
  popularityDesc = "popularity.desc",
  voteaverageAsc = "vote_average.asc",
  voteaverageDesc = "vote_average.desc",
  primaryreleasedateAsc = "primary_release_date.asc",
  primaryreleasedateDesc = "primary_release_date.desc",
  titleAsc = "title.asc",
  titleDesc = "title.desc",
}